requirements.txt
Click==6.7
Faker==0.9.0
PyYAML==3.13
py-IoticAgent==0.6.4
